# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "InstiApp API",
    "description": "",
    "version": ""
  },
  "host": "localhost",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/bodies": {
      "get": {
        "operationId": "bodies_list",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "parameters": [],
        "description": "Body",
        "summary": "Body",
        "tags": [
          "bodies"
        ]
      },
      "post": {
        "operationId": "bodies_create",
        "responses": {
          "201": {
            "description": ""
          }
        },
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "description": "",
                  "type": "string"
                },
                "short_description": {
                  "description": "",
                  "type": "string"
                },
                "description": {
                  "description": "",
                  "type": "string"
                },
                "image_url": {
                  "description": "",
                  "type": "string"
                },
                "website_url": {
                  "description": "",
                  "type": "string"
                },
                "blog_url": {
                  "description": "",
                  "type": "string"
                },
                "cover_url": {
                  "description": "",
                  "type": "string"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "description": "Create Body.\nNeeds the `AddB` Institute Role permission.",
        "summary": "Create Body.",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "bodies"
        ]
      }
    },
...
    "/api/bodies/{id}/follow": {
      "get": {
        "operationId": "bodies_follow",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "A UUID string identifying this Body.",
            "type": "string"
          }
        ],
        "description": "Follow or unfollow a body {?action}=0,1",
        "summary": "Follow or unfollow a body {?action}=0,1",
        "tags": [
          "bodies"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}